+++ RELEASE_NOTES.md
@@ -3,6 +3,7 @@
 * New `paket find-refs` command - http://fsprojects.github.io/Paket/paket-find-refs.html
 * If `paket add` doesn't change the dependencies file then the resolver will be skipped
 * BUGFIX: Trailing whitespace is ignored in `paket.references`
+* Migration of NuGet source credentials on convert-from-nuget - http://fsprojects.github.io/Paket/convert-from-nuget.html#Migrating-NuGet-source-credentials
 
 #### 0.12.2 - 07.11.2014
 * Adding --ignore-constraints to `paket outdated` - https://github.com/fsprojects/Paket/issues/308